refactor(tui): extract input mutation orchestration component
